Error in processing tuple order

+++ simple_example.cpp
@@ -2,7 +2,7 @@
 // This could be put into mex_wrap.cxx but this way allows naming of the mex function in matlab to this file name (without the .cxx)
 #include "mex_stuff.h"
 
-void mex_function(const double &x, const double &y, const double &z, double& result) {
+void mex_function(const double &x, const double &y, const double &z, int& result) {
 	result = (x + y)*z;
 }
